cleanup examples and call library
