+++ src/syslog.rs
@@ -25,7 +25,7 @@ const DEFAULT_FACILITY: &str = "user.*";
 /// * fail to open /etc/rsyslog.d/50-default.conf
 /// * fail to write modified contents to /etc/rsyslog.d/50-default.conf
 /// * fail to restart rsyslogd service
-pub fn set(remote_addrs: &Option<Vec<String>>) -> Result<()> {
+pub fn set(remote_addrs: &Option<Vec<String>>) -> Result<bool> {
     if let Some(addrs) = remote_addrs {
         for addr in addrs {
             let _addr = addr
