--- test-roxy/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "test_roxy"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0"
-base64 = "0.13"
-bincode = "1.3"
-chrono = "0.4"
-log = "0.4"
-serde = { version = "1",  features = ["derive"] }
-serde_json = "1"
